+++ flumine/backtest/simulated.py
@@ -400,32 +400,59 @@ class Simulated:
 
     @property
     def profit(self) -> float:
-        if self.order.runner_status == "WINNER":
-            number_of_dead_heat_winners = self.order.number_of_dead_heat_winners or 1
-            profit = (self.size_matched / number_of_dead_heat_winners) * (
-                self.average_price_matched - 1
+        number_of_dead_heat_winners = self.order.number_of_dead_heat_winners or 1
+        if self.order.market_type == "EACH_WAY":
+            divisor = self.order.each_way_divisor
+            win = self.size_matched * (self.average_price_matched - 1)
+            place = self.size_matched * (
+                (self.average_price_matched - 1) * (1 / divisor)
             )
-
-            if number_of_dead_heat_winners == 2:
-                profit = profit - (self.size_matched / number_of_dead_heat_winners)
-            elif number_of_dead_heat_winners > 2:
-                profit = profit - (
-                    self.size_matched
-                    * (number_of_dead_heat_winners - 1)
-                    / number_of_dead_heat_winners
-                )
-
-            if self.side == "LAY":
-                profit = -profit
-
-            return round(profit, 2)
-        elif self.order.runner_status == "LOSER":
-            if self.side == "BACK":
-                return -self.size_matched
+            if number_of_dead_heat_winners > 1:  # todo dead heat
+                logger.error("Each Way Dead Heat Detected but not handled")
+            if self.order.runner_status == "WINNER":
+                profit = round(win + place, 2)
+                if self.side == "BACK":
+                    return profit
+                else:
+                    return -profit
+            elif self.order.runner_status == "PLACED":
+                if self.side == "BACK":
+                    profit = place - self.size_matched
+                    return round(profit, 2)
+                else:
+                    profit = self.size_matched - place
+                    return round(profit, 2)
+            elif self.order.runner_status == "LOSER":
+                matched = round(self.size_matched * 2, 2)
+                if self.side == "BACK":
+                    return -matched
+                else:
+                    return matched
             else:
-                return self.size_matched
+                return 0.0
         else:
-            return 0.0
+            if self.order.runner_status == "WINNER":
+                profit = (self.size_matched / number_of_dead_heat_winners) * (
+                    self.average_price_matched - 1
+                )
+                if number_of_dead_heat_winners == 2:
+                    profit = profit - (self.size_matched / number_of_dead_heat_winners)
+                elif number_of_dead_heat_winners > 2:
+                    profit = profit - (
+                        self.size_matched
+                        * (number_of_dead_heat_winners - 1)
+                        / number_of_dead_heat_winners
+                    )
+                if self.side == "LAY":
+                    profit = -profit
+                return round(profit, 2)
+            elif self.order.runner_status == "LOSER":
+                if self.side == "BACK":
+                    return -self.size_matched
+                else:
+                    return self.size_matched
+            else:
+                return 0.0
 
     @property
     def status(self) -> str: